Blit works with floats

+++ Project/Dev_class11_handout/Motor2D/j1Window.cpp
@@ -41,7 +41,7 @@ bool j1Window::Awake(pugi::xml_node& config)
 		width = config.child("resolution").attribute("width").as_int(640);
 		height = config.child("resolution").attribute("height").as_int(480);
 		scale = config.child("resolution").attribute("scale").as_int(1);
-		scale = 2;
+		scale = 2.0f;
 		if(fullscreen == true)
 		{
 			flags |= SDL_WINDOW_FULLSCREEN;
@@ -108,7 +108,7 @@ void j1Window::GetWindowSize(uint& width, uint& height) const
 	height = this->height;
 }
 
-uint j1Window::GetScale() const
+float j1Window::GetScale() const
 {
 	return scale;
 }
